requirements.txt
blinker==1.4
boto==2.49.0
boto3==1.17.44
botocore==1.20.44
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
idna==2.10
invoke==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.2
paypalhttp==1.0.0
pika==1.2.0
protobuf==3.15.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.4
Werkzeug==1.0.1